grep tests: prepare for testing regexes containing binary data
Add setup code needed for testing regexes that contain binary
data. The regcomp() function can't support that, because it takes a
\0-delimited char*, but other regex such as PCRE v2 are capable of
taking a pattern with a length.

In order to test that patterns like that can properly match as regexes
and not just as fixed string matches the test content needs to have
some regex metacharacters in it.

Add a few of those, this doesn't change the semantics of any of the
existing tests.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>