grep: prepare for testing binary regexes containing rx metacharacters
Add setup code needed for testing regexes that contain both binary
data and regex metacharacters. The regcomp() function can't support
that, because it takes a \0-delimited char*, but other regex such as
PCRE v2 are capable of taking a pattern with a length.

When kwset was imported in commit 9eceddeec6 ("Use kwset in grep",
2011-08-21) this limitation was fixed, but it introduced the
undocumented limitation that any pattern containing \0 implicitly
becomes a fixed string.

That's not necessarily something we'd like to keep in the future,
e.g. PCRE v2 is capable of parsing a regex that contains \0.

In order to test that patterns like that can properly match as regexes
and not just as fixed string matches the test content needs to have
some regex metacharacters in it.

There were other blind spots in the tests. The code around kwset
specially handles case-insensitive & non-ASCII data, but there were no
tests for this.

Fix all of that by amending the text being matched to contain both
regex metacharacters & non-ASCII data.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>