tests: remove the GETTEXT_POISON compile-time option to test i18n marking
Remove the GETTEXT_POISON=YesPlease compile-time which turns all of
git's LC_*=C output into strings like "# GETTEXT POISON #" instead of
gettext(msgid).

See commit bb946bba76 ("i18n: add GETTEXT_POISON to simulate
unfriendly translator", 2011-02-22) for what this was originally
intended for.

This facility has been broken for quite a while and has been subjected
to frequent bitrot. The initial idea behind it back when it was added
in 2011 was to prevent the accidental translation of plumbing
messages.

This isn't a big concern anymore as git isn't mass-adding i18n
messages for a newly developed i18n facility as it was back then,
maintaining this facility incurs a burden, and in actuality this has
often been broken long enough for potential plumbing messages to be
translated & make their way into major releases anyway.

Most of this patch consists of search/replacing the test suite for:

    test_i18ngrep ! -> ! grep
    test_i18ngrep   -> grep
    test_i18ncmp    -> test_cmp

1. <AANLkTi=5MrU-JyeQ3UVNbVwzn-8FbstUXafgcQaLWXDB@mail.gmail.com>
   (https://public-inbox.org/git/AANLkTi=5MrU-JyeQ3UVNbVwzn-8FbstUXafgcQaLWXDB@mail.gmail.com/)

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>