test: remove unused parameter from the wildmatch test
Remove the unused second parameter to the match() function. Since
commit 70a8fc999d ("stop using fnmatch (either native or compat)",
2014-02-15) this hasn't been used, but was previously used to indicate
ok/fail if the fnmatch() backend was in use.

The "x" argument meant that the test would be skipped for
fnmatch(). Support for that was implicitly added in commit
feabcc173b ("Integrate wildmatch to git", 2012-10-15) by leaving out
an "else" condition, and made use of starting in commit
b79c0c3755 ("wildmatch: properly fold case everywhere", 2013-05-30).

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>