t/perf: add a GIT_PERF_MAKE_COMMAND for when *_MAKE_OPTS is too limited
Add a git GIT_PERF_MAKE_COMMAND variable to compliment the existing
GIT_PERF_MAKE_OPTS facility. This allows specifying an arbitrary
shellscript to execute instead of 'make'.

This is useful in cases where e.g. the name, semantics or defaults of
a Makefile flag have changed over time. It can even be used to change
the contents of the tree, useful for monkeypatching ancient versions
of git to get them to build.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>