diffcore-pickaxe: abort early when -G finds a match
Change the diff_grep() callback logic so that when -G finds a match
the diffcore aborts early. Since commit f506b8e8b5 ("git log/diff: add
-G<regexp> that greps in the patch text", 2010-08-23) there's been a
"NEEDSWORK" comment indicating that this should be done.

Before this the code will find a match in the version A & B of the
file, then the diff machinery will be invoked to generate a diff for
the entire file. When a hit is found the callback just exits early,
but the entire diff is still generated.

Now through a massive hack we abort the entire diff generation
early. This works, but should ideally be refactored to be so that
this:

    typedef void (*xdiff_emit_consume_fn)(void *, char *, unsigned long);

Can optionally return some "abort early" code instead of having a void
return signature.

This doesn't seem to do much, on git.git:

    $ GIT_PERF_REPEAT_COUNT=3 GIT_PERF_REPO=~/g/git GIT_PERF_MAKE_OPTS='-j8 CFLAGS=-O3' ./run v2.13.0 HEAD  -- p4210-log-pickaxe.sh
    [...]
    Test                                      v2.13.0             HEAD
    -------------------------------------------------------------------------------------
    4210.1: log -G'a.*b.*c'                   12.92(12.68+0.21)   12.88(12.65+0.20) -0.3%
    4210.2: log -G'how.to'                    12.65(12.50+0.12)   12.57(12.43+0.11) -0.6%
    4210.3: log -G'^how to'                   12.71(12.57+0.11)   12.62(12.52+0.08) -0.7%
    4210.4: log -G'[how] to'                  12.80(12.65+0.12)   12.72(12.54+0.15) -0.6%
    4210.5: log -G'\(e.t[^ ]*\|v.ry\) rare'   12.69(12.52+0.12)   12.62(12.50+0.09) -0.6%
    4210.6: log -G'm\(ú\|u\)lt.b\(æ\|y\)te'   12.81(12.69+0.10)   12.62(12.52+0.07) -1.5%

Here all of the confidence intervals overlap. I thought maybe this had
a bigger effect on a repo like gcc which has lots of very long files,
here's gcc.git against just the ChangeLog file:

    Test                                      v2.13.0           HEAD
    ---------------------------------------------------------------------------------
    4210.1: log -G'a.*b.*c'                   6.16(5.73+0.38)   6.13(5.74+0.36) -0.5%
    4210.2: log -G'how.to'                    6.09(5.76+0.30)   6.04(5.69+0.33) -0.8%
    4210.3: log -G'^how to'                   6.07(5.64+0.40)   6.07(5.68+0.36) +0.0%
    4210.4: log -G'[how] to'                  6.08(5.64+0.38)   6.10(5.72+0.34) +0.3%
    4210.5: log -G'\(e.t[^ ]*\|v.ry\) rare'   6.07(5.71+0.33)   6.08(5.77+0.30) +0.2%
    4210.6: log -G'm\(ú\|u\)lt.b\(æ\|y\)te'   6.08(5.67+0.38)   6.06(5.64+0.39) -0.3%

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>