sha1dc: use sha1collisiondetection as a submodule
Replace the forked sha1dc directory with a copy of the upstream code
imported as a submodule. This is the exact same code as now exists in
the sha1dc/ directory.

The initial reason for copy/pasting the code into sha1dc and locally
modifying it was that it needed to be altered to work with the git
project.

The upstream project has accepted my code changes to allow us to use
their code as-is, see the preceding commit for details. So import the
code as a submodule instead, this will make it easier to keep
up-to-date with any upstream fixes or improvements.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>